% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_difference.R
\name{calculate_difference}
\alias{calculate_difference}
\title{Calculate Differences in Omics Data Between Visits}
\usage{
calculate_difference(omics_data, patient_data)
}
\arguments{
\item{omics_data}{A \link{data.frame} that contains the omics data with
row names as sample IDs.}

\item{patient_data}{A \link{data.frame} that contains the patient data with
columns for fxs_sts_id, Visit1, and Visit3.}
}
\value{
A \link{data.frame} containing the differences (Visti 3 - Visit 1) in
omics data for each patient.
}
\description{
The Similarity Network Fusion does not handle longitudinal data, so this
function calculates the difference between Visit 3 and Visit 1 for each
patient prior to fitting SNF.
}
